pyproject.toml
[project]
name = "qa-automation-project"
version = "0.1.0"
description = "QA Automation project for UI and API testing."
authors = [
    { name = "Slava", email = "yaroslavpimenov92@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pytest==8.2.2",
    "pytest-html==4.1.1",
    "configparser==7.0.0",
    "selenium==4.22.0",
    "webdriver-manager==4.0.2",
    "faker==37.5.3",
    "allure-pytest==2.15.0",
    "tenacity==9.1.2",
    "httpx==0.28.1",
    "python-dotenv==1.1.1"
]


[tool.uv]